refactor(oracle): migrate arp to per-protocol plugins across stages
